feature/multi-model: added support for multi model test
